requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
pytest==8.3.3